requirements.txt
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.4.0
tabulate==0.8.10
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
